refactor(Shows): ui
move "Shows" component to ui folder
