+++ src/pages/Shows/ui/Shows.jsx
@@ -1,6 +1,6 @@
 import 'src/styles/shows.scss'
-import {useShowsPageViewModel} from "./model";
-import {ShowesPagePure} from "./ui/components";
+import {useShowsPageViewModel} from "../model";
+import {ShowesPagePure} from "./components";
 
 const Shows = () => {
     const {setPageNumber, sliced, data, pageNumber, isLoading} = useShowsPageViewModel();
